Make MeshDisintegrator crash less while seeking animation in the editor
Comment from the code:
Update2: Going back to "deferred", since there was strange occasional crashes (at the same time I also changed the sourceTextMesh.duplicates to duplicate also subresources, so don't sure if that was the actual reason for crashes (based on the strange font-signal error prints it probably was, though...)).

Anyway... After twiddling with the editor I couldn't get crashes to happen, so maybe it's ok now.