Miscellaneous tunings, also migrate to Godot 4.1
These are changes made during the last 3 months or so. I can't remember what has changed to be honest (and due to animation tracks losing their ordering it's hard to figure out from diffs). At least Godot 4.1 (and 4.1.1) appeared after the last commit, though.

After the last commit and most of these (then uncommitted) miscellaneous changes I dived into another project for a while and after that lost all motivation to get this any further. Tried few times to get on with this, always ending up doomscrolling or something. Maybe doomscrolling is consistent with the theme of this, though... Need to stop it to get something done.