Fix satellite's commonAntennaBaseLocalTranslation
Don't know exactly why and when this broke. Apparently the translation needs to be 10 times what it was. Probably this is related to antennas being scaled by 10 due to CSG working better on larger scales. Maybe some bugfix broke this.

Also metallic texture channel was converted automatically when touching the scene.