Throw unnecessary junk away from SubViewport_Elite.gd
Changing the resolution of the subViewport so that every pixel corresponds to some integer multiple of screen pixels didn't work out too well (or I was too stupid to get it working). The code cut here is related to these. This idea was replaced with custom anti-aliasing trying to preserve the blocky look of Amiga-Elite.