Replace fakeDistanceOverride with "follower"
Couldn't get the previous try working to fake the world (planet at the these distances) being infinitely far.

So why this? Well, the space station is like 300 m in diameter and the planet is only 2 km away. Flying around the station breaks the illusion of "spacial vastness" totally when the planet moves in relation to the sky. Now it should look like it stays at the same place (when it's actually moving with the camera).

Some other tunings also.