Fix weird nag
Don't know what happened but after trying to edit/run this project on the newer versions of Godot (and after deleting .godot-directory), loading of the Global.gd failed. Something related to OscilloscopeDataStorage being of unknown type or something and then Global.gd not extending Node (or something, don't remember exactly).