Add (commented out) debug prints with tics_msec to almost all _ready-functions
Did this to make it possible to profile where the "loading"-time is spent. At least a lot of solar cells are created so these may be use somewhat large portion of the loading time. But creating those is quite difficult to refactor, so not doing it now.

Should probably create some heaviest sub-scenes in separate threads to make the main thread responsive at the start (instead of "not responding").

Also printing everything also slows down the "loading" so measurements are not too accurate.