Add shader precompiling
This is now done so that defined points in time in animation will be played for some time (animation paused). These points in time are selected so that they represent all materials shown during the selected animation.

Also some animation tuning was needed to fix values remembered from the end of the animation as it is now "fast forwarded" when recompiling the shaders.

TODO: Define "points in time" for the happier animation