Workaround for instance shader parameter export bug(?)
It seems that these instance shader parameters do not make their way to exported project(?) In this case this causes Thargoid debris field to disappear. Since only exploding "Elite"-ship is Thargoid, putting the appropriate values straight into the shader source.

Didn't find a bug report and/or fix about this, so maybe should dig deeper and make one if this is still relevant...