Optimize solar cell and panel creation
Solar cell was a scene before and that scene was created like 1769 times when starting the project. Not very heavy scene, but this amount takes some time anyway. Now the scene is replaced with a local node that is duplicated when creating the "static" solar panels ("constructive" panels seen in the beginning of the demo).

The scene is also still used in some part (debris). The script is now removed from these as it doesn't give any benefit now (don't remember why there was definable (@export) materials there, probably for some prototyping purposes...)