Add a plane mesh in front of the space station to draw Thargoid debris into
This ensures that all of it is drawn. Previous version with a sphere mesh caused some parts of the debris to be drawn behind the station because the near part of the sphere was cut by cameras near field or was behind the camera altogether. Far side was behind some parts of the station (bottom rim at least).