Tune JellyPlanet glow and drawing order
Jelly planet vertices are now "pushed back" by some amount while ascending to space station. Center glow (in the center of the "infinite zoomer") is now adjusted on the fly. Had to make an ugly hack to keep glow working and "push back" the vertices. Tried to work with local vertex coordinates, but just couldn't get pushing back to work. So switched to world_vertex_coords and this broke the center glow adjustment that is now handled using UVs.

May need to add pushBackFactor and groundZeroGlowFactor to some init anims. Lets see...